# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MidRequest": {
        "properties": {
          "mid": {
            "minimum": 1.0,
            "title": "Mid",
            "type": "integer"
          }
        },
        "required": [
          "mid"
        ],
        "title": "MidRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Open-source self-hosted toolkit for inspecting and cleaning your own Bilibili account: followings, favorite folders, dynamics, and watch history.\n\nAll requests are rate-limited (default 1.5 r/s, shared per event loop) and retried with exponential backoff on risk-control responses (`-352`, `-799`, `-509`, HTTP 412/429) — 3 retries, 4 attempts total. Long-running cleans are exposed as async tasks under `/api/v2/tasks/*`.\n\nEndpoints under `/api/v2/*` are the canonical AI-facing surface; `/api/clean/*` (v1) are preserved aliases.\n\nDeletes are permanent — Bilibili has no undo. Only the logged-in account can be operated on.",
    "title": "Bilibili Cleaner",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/clean/all": {
      "post": {
        "operationId": "clean_all_api_clean_all_post",
        "parameters": [
          {
            "in": "header",
            "name": "SESSDATA",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sessdata"
            }
          },
          {
            "in": "header",
            "name": "bili-jct",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bili Jct"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MidRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clean All Api Clean All Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean All",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/clean/dynamics": {
      "post": {
        "operationId": "clean_dynamics_api_clean_dynamics_post",
        "parameters": [
          {
            "in": "header",
            "name": "SESSDATA",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sessdata"
            }
          },
          {
            "in": "header",
            "name": "bili-jct",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bili Jct"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MidRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clean Dynamics Api Clean Dynamics Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean Dynamics",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/clean/favorites": {
      "post": {
        "operationId": "clean_favorites_api_clean_favorites_post",
        "parameters": [
          {
            "in": "header",
            "name": "SESSDATA",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sessdata"
            }
          },
          {
            "in": "header",
            "name": "bili-jct",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bili Jct"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MidRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clean Favorites Api Clean Favorites Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean Favorites",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/clean/followings": {
      "post": {
        "operationId": "clean_followings_api_clean_followings_post",
        "parameters": [
          {
            "in": "header",
            "name": "SESSDATA",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sessdata"
            }
          },
          {
            "in": "header",
            "name": "bili-jct",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bili Jct"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MidRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clean Followings Api Clean Followings Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean Followings",
        "tags": [
          "v1"
        ]
      }
    },
    "/api/clean/history": {
      "post": {
        "operationId": "clean_history_api_clean_history_post",
        "parameters": [
          {
            "in": "header",
            "name": "SESSDATA",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sessdata"
            }
          },
          {
            "in": "header",
            "name": "bili-jct",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bili Jct"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clean History Api Clean History Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean History",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/qrcode/poll/{qrcode_key}": {
      "get": {
        "operationId": "poll_qrcode_api_qrcode_poll__qrcode_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "qrcode_key",
            "required": true,
            "schema": {
              "title": "Qrcode Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Poll Qrcode Api Qrcode Poll  Qrcode Key  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll Qrcode",
        "tags": [
          "v1"
        ]
      }
    },
...
  },
...
}